requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
clarifai==2.6.2
configparser==3.8.1
cryptography==2.8
emoji==0.5.4
future==0.18.2
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
jsonschema==2.6.0
protobuf==3.11.1
pycparser==2.19
PySocks==1.7.1
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.2.0
requests==2.22.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
dnspython==1.16.0
pymongo==3.10.1
